Updated Loan ROIs, Deleted Old Files
Updated gitignore to ignore new payments CSV files as they are both over 5 GB.

Recalculated the ROI of loans with errors.

Calculated ROI of loans that are new to the training set.

Deleted supplemental interest rate CSV files since they would not be available at the time predictions need to be made.